+++ src/Maw.Domain/Videos/IVideoService.cs
@@ -12,15 +12,15 @@ public interface IVideoService
     Task<Video?> GetVideoAsync(short id, string[] roles);
     Task<Category?> GetCategoryAsync(short categoryId, string[] roles);
     Task<IEnumerable<Comment>> GetCommentsAsync(short videoId, string[] roles);
-    Task<GpsDetail?> GetGpsDetailAsync(int videoId, string[] roles);
+    Task<GpsDetail?> GetGpsDetailAsync(short videoId, string[] roles);
     Task<Rating?> GetRatingsAsync(short videoId, string username, string[] roles);
     Task InsertCommentAsync(short videoId, string username, string comment, string[] roles);
     Task<float?> SaveRatingAsync(short videoId, string username, short rating, string[] roles);
     Task<float?> RemoveRatingAsync(short videoId, string username, string[] roles);
 
     // admin functions
-    Task SetGpsOverrideAsync(int videoId, GpsCoordinate gps, string username);
-    Task SetCategoryTeaserAsync(short categoryId, int videoId);
+    Task SetGpsOverrideAsync(short videoId, GpsCoordinate gps, string username);
+    Task SetCategoryTeaserAsync(short categoryId, short videoId);
     Task ClearCacheAsync();
     Task<IEnumerable<CategoryAndRoles>> GetCategoriesAndRolesAsync();
 }
